+++ src/base/controllers/url.go
@@ -4,6 +4,8 @@ import (
 	"base/services"
 	"fmt"
 
+	"github.com/astaxie/beego"
+
 	//"encoding/json"
 )
 
@@ -17,18 +19,16 @@ type UrlController struct {
  *	@todo 		参数验证, 封装返回
  */
 func (r *UrlController) GoShorten() {
-	fmt.Println("9999999999999999999999999", r.Tr("outputParams.PARAMSILLEGAL"))
-
 	//接受参数 json raw
-	rawDataBody := r.Ctx.Input.RequestBody
 	rawMetaHeader := r.Ctx.Input.Request.Header
-	for key, val := range rawMetaHeader {
-		fmt.Println(key, ": ", val[0])
-	}
+	rawDataBody := r.Ctx.Input.RequestBody
+
+	//记录参数日志
+	beego.Trace("入参body:" + string(rawDataBody))
 
 	//调用servcie方法, 将参数传递过去
 	var su services.Url
-	shorten, err := su.GoShorten(rawDataBody, rawMetaHeader)
+	shorten, err := su.GoShorten(rawMetaHeader, rawDataBody)
 
 	var data = make(map[string]interface{})
 	if err.Err != nil {
