+++ src/base/cores/inputParams.go
@@ -4,23 +4,36 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/validation"
+	"github.com/pquerna/ffjson/ffjson"
 )
+
+type MetaHeader struct {
+	Source    []string `valid:"Required"`
+	Version   []string `valid:"Required"`
+	SecretKey []string `valid:"Required"`
+	RequestID []string `valid:"Required"`
+	Token     []string `valid:"Required"`
+	IP        []string `valid:"IP"`
+}
 
 /**
  *	@auther		jream.lu
  *	@intro		入参验证
  *	@logic
  *	@todo		返回值
- *	@params		params ...interface{}	切片指针
+ *	@data		data ...interface{}	切片指针
  *	@return 	?
  */
-func InputParamsCheck(params ...interface{}) int {
-
+func InputParamsCheck(meta map[string][]string, data ...interface{}) int {
+	//DataParams check
 	valid := validation.Validation{}
 
-	for key, val := range params {
+	for key, val := range data {
 		is, err := valid.Valid(val)
+
+		//日志
 
 		if err != nil {
 			// handle error
@@ -33,9 +46,26 @@ func InputParamsCheck(params ...interface{}) int {
 			}
 		}
 	}
+
+	//MetaHeader check
+	MetaHeaderCheck(meta)
 	return 1
 }
 
-//meta参数验证
+/**
+ * meta参数验证
+ * 1.map转json
+ * 2.json转slice
+ * 3.解析到struct
+ */
+func MetaHeaderCheck(meta map[string][]string) int {
+	rawMetaHeader, _ := ffjson.Marshal(meta)
+	beego.Trace("入参meta:" + string(rawMetaHeader))
+	var metaHeader MetaHeader
+	ffjson.Unmarshal(rawMetaHeader, &metaHeader)
+	fmt.Println("meta json解析:", metaHeader)
+
+	return 1
+}
 
 //request id增加
